+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Pre-releases
+## v0.3.2-alpha
+* Front-end now parses PORT and HOST from config file as well
+* Server parses PORT from /etc/papayachat/server.config file
+* Server availability test automatically parses PORT from ./etc repo files
+* A unit test C file for configParser.c is also added to ./tests as test_configParser.c
 ## v0.3.1-alpha
 * [FIXED] client's config file is now handled within the home directory of the user
 * The greetings message delivered by the server after establishing a connection now,